Level-4. Added Event, Deadline and Todo subclasses of Task. Updated if-else in levelFourRespond to handle the 3 subclasses.
